requirements.txt
Babel==2.14.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.7
engineering-notation==0.10.0
idna==3.6
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
num2words==0.5.13
python-docx==1.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
tk-tools==0.16.0
typing-extensions==4.9.0
urllib3==2.1.0